refactor(hydro_lang)!: move `boundedness` module under `live_collections`
Type tags for boundedness are only used in the type parameters for live collections.
