+++ hydro_lang/src/live_collections/keyed_stream.rs
@@ -3,10 +3,10 @@ use std::marker::PhantomData;
 
 use stageleft::{IntoQuotedMut, QuotedWithContext, q};
 
+use super::boundedness::{Bounded, Boundedness, Unbounded};
 use super::keyed_singleton::KeyedSingleton;
 use super::optional::Optional;
 use super::stream::{ExactlyOnce, MinOrder, MinRetries, NoOrder, Stream, TotalOrder};
-use crate::boundedness::{Bounded, Boundedness, Unbounded};
 use crate::builder::ir::HydroNode;
 use crate::cycle::{CycleCollection, CycleComplete, ForwardRefMarker};
 use crate::location::tick::NoAtomic;
@@ -568,7 +568,7 @@ where
     /// A variant of [`Stream::fold`], intended for keyed streams. The aggregation is executed in-order across the values
     /// in each group. But the aggregation function returns a boolean, which when true indicates that the aggregated
     /// result is complete and can be released to downstream computation. Unlike [`Stream::fold_keyed`], this means that
-    /// even if the input stream is [`crate::boundedness::Unbounded`], the outputs of the fold can be processed like normal stream elements.
+    /// even if the input stream is [`super::boundedness::Unbounded`], the outputs of the fold can be processed like normal stream elements.
     ///
     /// # Example
     /// ```rust
