+++ hydro_lang/src/location/tick.rs
@@ -5,13 +5,13 @@ use sealed::sealed;
 use stageleft::{QuotedWithContext, q};
 
 use super::{Cluster, Location, LocationId, Process};
-use crate::boundedness::Bounded;
 use crate::builder::FlowState;
 use crate::builder::ir::{HydroNode, HydroSource};
 use crate::cycle::{
     CycleCollection, CycleCollectionWithInitial, DeferTick, ForwardRef, ForwardRefMarker,
     TickCycle, TickCycleMarker,
 };
+use crate::live_collections::boundedness::Bounded;
 use crate::live_collections::optional::Optional;
 use crate::live_collections::singleton::Singleton;
 use crate::live_collections::stream::{ExactlyOnce, Stream, TotalOrder};
